REFACTOR: Move algebraic interface for wiring diagrams into submodule.
The core submodule for wiring diagams was overcrowded.
