+++ src/wiring_diagrams/WiringDiagrams.jl
@@ -4,6 +4,7 @@ using Reexport
 
 include("Core.jl")
 include("Layers.jl")
+include("Algebraic.jl")
 include("Algorithms.jl")
 include("Expressions.jl")
 include("Serialization.jl")
@@ -12,12 +13,12 @@ include("JSON.jl")
 
 @reexport using .WiringDiagramCore
 @reexport using .WiringLayers
+@reexport using .AlgebraicWiringDiagrams
 @reexport using .WiringDiagramAlgorithms
 @reexport using .WiringDiagramExpressions
 
 using .WiringDiagramSerialization
 export convert_from_graph_data, convert_to_graph_data
-
 @reexport using .GraphMLWiringDiagrams
 @reexport using .JSONWiringDiagrams
